requirements.txt
amqp==5.2.0
asgiref==3.10.0
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt==3.2.2
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloud-init==25.1.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
dbus-python==1.3.2
distro==1.9.0
distro-info==1.7+build1
Django==5.2.7
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.2
django-js-asset==3.1.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
environ==1.0
ephem==4.1.5
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kombu==5.3.5
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
Markdown==3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
netifaces==0.11.0
oauthlib==3.2.2
pillow==12.0.0
prompt-toolkit==3.0.43
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycurl==7.45.3
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5
python-dateutil==2.8.2
python-memcached==1.58
pytz==2024.1
PyYAML==6.0.1
redis==7.0.1
requests==2.31.0
rich==13.7.1
service-identity==24.1.0
setuptools==68.1.2
six==1.16.0
sqlparse==0.5.3
systemd-python==235
Twisted==24.3.0
typing_extensions==4.10.0
tzdata==2025.2
ubuntu-pro-client==8001
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.0.7
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.5
wheel==0.42.0
whitenoise==6.11.0
zope.interface==6.1